+++ hcloud/schema.go
@@ -514,8 +514,20 @@ func LoadBalancerTargetFromSchema(s schema.LoadBalancerTarget) LoadBalancerTarge
 			Server: &Server{ID: s.Server.ID},
 		}
 	}
+	if s.LabelSelector != nil {
+		lt.LabelSelector = &LoadBalancerTargetLabelSelector{
+			Selector: s.LabelSelector.Selector,
+		}
+	}
+	if s.IP != nil {
+		lt.IP = &LoadBalancerTargetIP{IP: s.IP.IP}
+	}
+
 	for _, healthStatus := range s.HealthStatus {
 		lt.HealthStatus = append(lt.HealthStatus, LoadBalancerTargetHealthStatusFromSchema(healthStatus))
+	}
+	for _, target := range s.Targets {
+		lt.Targets = append(lt.Targets, LoadBalancerTargetFromSchema(target))
 	}
 	return lt
 }
@@ -709,6 +721,12 @@ func loadBalancerCreateOptsToSchema(opts LoadBalancerCreateOpts) schema.LoadBala
 		case LoadBalancerTargetTypeServer:
 			schemaTarget.Type = string(LoadBalancerTargetTypeServer)
 			schemaTarget.Server = &schema.LoadBalancerCreateRequestTargetServer{ID: target.Server.Server.ID}
+		case LoadBalancerTargetTypeLabelSelector:
+			schemaTarget.Type = string(LoadBalancerTargetTypeLabelSelector)
+			schemaTarget.LabelSelector = &schema.LoadBalancerCreateRequestTargetLabelSelector{Selector: target.LabelSelector.Selector}
+		case LoadBalancerTargetTypeIP:
+			schemaTarget.Type = string(LoadBalancerTargetTypeIP)
+			schemaTarget.IP = &schema.LoadBalancerCreateRequestTargetIP{IP: target.IP.IP}
 		}
 		req.Targets = append(req.Targets, schemaTarget)
 	}