+++ hcloud/schema/load_balancer.go
@@ -82,10 +82,13 @@ type LoadBalancerServiceHealthCheckHTTP struct {
 }
 
 type LoadBalancerTarget struct {
-	Type         string                           `json:"type"`
-	Server       *LoadBalancerTargetServer        `json:"server"`
-	HealthStatus []LoadBalancerTargetHealthStatus `json:"health_status"`
-	UsePrivateIP bool                             `json:"use_private_ip"`
+	Type          string                           `json:"type"`
+	Server        *LoadBalancerTargetServer        `json:"server"`
+	LabelSelector *LoadBalancerTargetLabelSelector `json:"label_selector"`
+	IP            *LoadBalancerTargetIP            `json:"ip"`
+	HealthStatus  []LoadBalancerTargetHealthStatus `json:"health_status"`
+	UsePrivateIP  bool                             `json:"use_private_ip"`
+	Targets       []LoadBalancerTarget             `json:"targets,omitempty"`
 }
 
 type LoadBalancerTargetHealthStatus struct {
@@ -97,6 +100,14 @@ type LoadBalancerTargetServer struct {
 	ID int `json:"id"`
 }
 
+type LoadBalancerTargetLabelSelector struct {
+	Selector string `json:"selector"`
+}
+
+type LoadBalancerTargetIP struct {
+	IP string `json:"ip"`
+}
+
 type LoadBalancerListResponse struct {
 	LoadBalancers []LoadBalancer `json:"load_balancers"`
 }
@@ -106,13 +117,23 @@ type LoadBalancerGetResponse struct {
 }
 
 type LoadBalancerActionAddTargetRequest struct {
-	Type         string                                    `json:"type"`
-	Server       *LoadBalancerActionAddTargetRequestServer `json:"server,omitempty"`
-	UsePrivateIP *bool                                     `json:"use_private_ip,omitempty"`
+	Type          string                                           `json:"type"`
+	Server        *LoadBalancerActionAddTargetRequestServer        `json:"server,omitempty"`
+	LabelSelector *LoadBalancerActionAddTargetRequestLabelSelector `json:"label_selector,omitempty"`
+	IP            *LoadBalancerActionAddTargetRequestIP            `json:"ip,omitempty"`
+	UsePrivateIP  *bool                                            `json:"use_private_ip,omitempty"`
 }
 
 type LoadBalancerActionAddTargetRequestServer struct {
 	ID int `json:"id"`
+}
+
+type LoadBalancerActionAddTargetRequestLabelSelector struct {
+	Selector string `json:"selector"`
+}
+
+type LoadBalancerActionAddTargetRequestIP struct {
+	IP string `json:"ip"`
 }
 
 type LoadBalancerActionAddTargetResponse struct {
@@ -120,12 +141,22 @@ type LoadBalancerActionAddTargetResponse struct {
 }
 
 type LoadBalancerActionRemoveTargetRequest struct {
-	Type   string                                       `json:"type"`
-	Server *LoadBalancerActionRemoveTargetRequestServer `json:"server,omitempty"`
+	Type          string                                              `json:"type"`
+	Server        *LoadBalancerActionRemoveTargetRequestServer        `json:"server,omitempty"`
+	LabelSelector *LoadBalancerActionRemoveTargetRequestLabelSelector `json:"label_selector,omitempty"`
+	IP            *LoadBalancerActionRemoveTargetRequestIP            `json:"ip,omitempty"`
 }
 
 type LoadBalancerActionRemoveTargetRequestServer struct {
 	ID int `json:"id"`
+}
+
+type LoadBalancerActionRemoveTargetRequestLabelSelector struct {
+	Selector string `json:"selector"`
+}
+
+type LoadBalancerActionRemoveTargetRequestIP struct {
+	IP string `json:"ip"`
 }
 
 type LoadBalancerActionRemoveTargetResponse struct {
@@ -234,13 +265,23 @@ type LoadBalancerCreateRequestAlgorithm struct {
 }
 
 type LoadBalancerCreateRequestTarget struct {
-	Type         string                                 `json:"type"`
-	Server       *LoadBalancerCreateRequestTargetServer `json:"server,omitempty"`
-	UsePrivateIP *bool                                  `json:"use_private_ip,omitempty"`
+	Type          string                                        `json:"type"`
+	Server        *LoadBalancerCreateRequestTargetServer        `json:"server,omitempty"`
+	LabelSelector *LoadBalancerCreateRequestTargetLabelSelector `json:"label_selector,omitempty"`
+	IP            *LoadBalancerCreateRequestTargetIP            `json:"ip,omitempty"`
+	UsePrivateIP  *bool                                         `json:"use_private_ip,omitempty"`
 }
 
 type LoadBalancerCreateRequestTargetServer struct {
 	ID int `json:"id"`
+}
+
+type LoadBalancerCreateRequestTargetLabelSelector struct {
+	Selector string `json:"selector"`
+}
+
+type LoadBalancerCreateRequestTargetIP struct {
+	IP string `json:"ip"`
 }
 
 type LoadBalancerCreateRequestService struct {