+++ hcloud/schema_test.go
@@ -1317,6 +1317,27 @@ func TestLoadBalancerFromSchema(t *testing.T) {
 					}
 				],
 				"use_private_ip": false
+			},
+			{
+				"type": "label_selector",
+				"label_selector": {
+					"selector": "lbt"
+				},
+				"targets": [
+					{
+						"type": "server",
+						"server": {
+							"id": 80
+						},
+						"health_status": [
+							{
+								"listen_port": 443,
+								"status": "healthy"
+							}
+						],
+						"use_private_ip": false
+					}
+				]
 			}
 		],
 		"algorithm": {
@@ -1365,7 +1386,7 @@ func TestLoadBalancerFromSchema(t *testing.T) {
 	if len(loadBalancer.Services) != 1 {
 		t.Errorf("unexpected length of Services: %v", len(loadBalancer.Services))
 	}
-	if len(loadBalancer.Targets) != 1 {
+	if len(loadBalancer.Targets) != 2 {
 		t.Errorf("unexpected length of Targets: %v", len(loadBalancer.Targets))
 	}
 	if loadBalancer.Algorithm.Type != "round_robin" {
@@ -1511,6 +1532,9 @@ func TestLoadBalancerTargetFromSchema(t *testing.T) {
 		if loadBalancerTarget.Server == nil || loadBalancerTarget.Server.Server.ID != 80 {
 			t.Errorf("unexpected Server: %v", loadBalancerTarget.Server)
 		}
+		if loadBalancerTarget.LabelSelector != nil {
+			t.Errorf("unexpected LabelSelector.Selector: %v", loadBalancerTarget.LabelSelector)
+		}
 		if loadBalancerTarget.UsePrivateIP {
 			t.Errorf("unexpected UsePrivateIP: %v", loadBalancerTarget.UsePrivateIP)
 		}
@@ -1523,6 +1547,80 @@ func TestLoadBalancerTargetFromSchema(t *testing.T) {
 			if loadBalancerTarget.HealthStatus[0].Status != LoadBalancerTargetHealthStatusStatusHealthy {
 				t.Errorf("unexpected HealthStatus[0].Status: %v", loadBalancerTarget.HealthStatus[0].Status)
 			}
+		}
+	})
+	t.Run("label_selector target", func(t *testing.T) {
+		data := []byte(`{
+			"type": "label_selector",
+			"label_selector": {
+				"selector": "lbt"
+			},
+			"targets": [
+				{
+					"type": "server",
+					"server": {
+						"id": 80
+					},
+					"health_status": [
+						{
+							"listen_port": 443,
+							"status": "healthy"
+						}
+					]
+				}
+			]
+		}`)
+		var s schema.LoadBalancerTarget
+		if err := json.Unmarshal(data, &s); err != nil {
+			t.Fatal(err)
+		}
+		loadBalancerTarget := LoadBalancerTargetFromSchema(s)
+		if loadBalancerTarget.Type != "label_selector" {
+			t.Errorf("unexpected Type: %v", loadBalancerTarget.Type)
+		}
+		if loadBalancerTarget.LabelSelector == nil || loadBalancerTarget.LabelSelector.Selector != "lbt" {
+			t.Errorf("unexpected LabelSelector: %v", loadBalancerTarget.LabelSelector)
+		}
+		if loadBalancerTarget.Server != nil {
+			t.Errorf("unexpected LabelSelector.Server: %v", loadBalancerTarget.Server)
+		}
+		if len(loadBalancerTarget.Targets) != 1 {
+			t.Errorf("unexpected Targets length: %v", len(loadBalancerTarget.Targets))
+		} else {
+			if loadBalancerTarget.Targets[0].Server == nil || loadBalancerTarget.Targets[0].Server.Server.ID != 80 {
+				t.Errorf("unexpected loadBalancerTarget.Targets[0].Server.Server.ID: %v", loadBalancerTarget.Targets[0].Server.Server.ID)
+			}
+			if len(loadBalancerTarget.Targets[0].HealthStatus) != 1 {
+				t.Errorf("unexpected Targets length: %v", len(loadBalancerTarget.Targets[0].HealthStatus))
+			} else {
+				if loadBalancerTarget.Targets[0].HealthStatus[0].ListenPort != 443 {
+					t.Errorf("unexpected HealthStatus[0].ListenPort: %v", loadBalancerTarget.Targets[0].HealthStatus[0].ListenPort)
+				}
+				if loadBalancerTarget.Targets[0].HealthStatus[0].Status != LoadBalancerTargetHealthStatusStatusHealthy {
+					t.Errorf("unexpected HealthStatus[0].Status: %v", loadBalancerTarget.Targets[0].HealthStatus[0].Status)
+				}
+			}
+		}
+	})
+
+	t.Run("ip target", func(t *testing.T) {
+		var s schema.LoadBalancerTarget
+
+		data := []byte(`{
+			"type": "ip",
+			"ip": {
+				"ip": "1.2.3.4"
+			}
+		}`)
+		if err := json.Unmarshal(data, &s); err != nil {
+			t.Fatal(err)
+		}
+		lbTgt := LoadBalancerTargetFromSchema(s)
+		if lbTgt.Type != LoadBalancerTargetTypeIP {
+			t.Errorf("unexpected Type: %s", lbTgt.Type)
+		}
+		if lbTgt.IP.IP != "1.2.3.4" {
+			t.Errorf("unexpected IP: %s", lbTgt.IP.IP)
 		}
 	})
 }
@@ -1860,6 +1958,10 @@ func TestLoadBalancerCreateOptsToSchema(t *testing.T) {
 							Server: &Server{ID: 5},
 						},
 					},
+					{
+						Type: LoadBalancerTargetTypeIP,
+						IP:   LoadBalancerCreateOptsTargetIP{IP: "1.2.3.4"},
+					},
 				},
 			},
 			Request: schema.LoadBalancerCreateRequest{
@@ -1908,6 +2010,12 @@ func TestLoadBalancerCreateOptsToSchema(t *testing.T) {
 						Type: "server",
 						Server: &schema.LoadBalancerCreateRequestTargetServer{
 							ID: 5,
+						},
+					},
+					{
+						Type: "ip",
+						IP: &schema.LoadBalancerCreateRequestTargetIP{
+							IP: "1.2.3.4",
 						},
 					},
 				},